jointcal tests on simulated data!
Cleanup print statements and size checks in jointcalTask.

Compute rms of 2 and 10 catalog runs as my test metric. It's not perfect, but
at least it provides a way to tell if I've broken anything. Still keeping around
the plotting code, as it's very useful for diagnostics.

Add validation_data_jointcal as optional dependency for the test calexps.