Get jointcal to work with new astrometrySourceSelector.
Cleanup config fields, removing some, and moving others to the old
starSelectorConfig, to be deleted later.

Add astrometrySourceSelector and use the registry to configure it.

Pull ccdImageList construction into separate method, with commented block
containing the old starSelector in case we need to reference it in the near
future. Also wrapped that with a profiling block, to help with speed tests.

Rename assoc -> associations

Tweak minSnr to keep relative/absolute RMS thresholds happy.