Work towards a passing parseAndRun test.
Had to deal with some of the metadata problems along the way, including adding
a fake calexp and detector to my FakeRef (which can entirely go away once we
fully solve the metadata thing) and adding an "LSST" branch with MJD-OBS to
CcdImage (lsstSim's camera is just called LSST).

Needed to reinstate coaddName to the config for parseAndRun.