JointcalStatistics bugfix re: visit tracking, add lsstSim pa1 test
Fixed a serious bug in how I was tracking visits in JointcalStatistics, as well
as a few less serious but not-trivial related bugs. In the process, cleaned up
the args that various methods take (removing ones that should just be computed
and handled internally) and renamed args in the plotting functions.

Added pa1 test to test_jointcal_lsstSim.py, using a value just above the current
computed value.

Several other docstring and comment cleanups.

Tweaked default plot_jointcal directory and mkdir if it's missing.