Big speed improvement and cleanup dead ccdImage code
Only read the calexps when building the ccdImages (and not when
building the output to write) and save the Calib to its respective
ccdImage to use it when preparing output. Use that calib in ccdImage
to convert flux to magnitude, instead of custom code.

No need for the calexp_md now: just extract the values from the calexp
(eventually this should come from some metadata datasetType, e.g.
`get('info')`).

Add SOURCE_IO_NO_FOOTPRINTS flag to get('src') for a big speedup,
since I don't need the heavy footprints (thanks @TallJimbo for the
suggestion).

Add individual profiling for each slow step, activated via
profile_jointcal option.

Remove a bunch of variables and getters from ccdImage that are totally
unused.