Add GAIA refcat and no-photo/no-astro HSC tests
Add hsc config for filterMap and tweak test to work with it.

Add every star number and chi2 metric for every test dataset.

Metrics tests always occur in JointcalTestBase._runJointcalTask(). Default
to checking every metric that gets returned, so they'll fail if that metric
isn't defined by the test method. A metric can be skipped by setting it to None.