Fix chi2 numbers for new model
PhotometryModel now operates on instFlux and instFluxErr, so the chi2 values
have moved around a little.

Fix bug in test_jointcal_cfht

Fix hsc no-WCS test to assert, as the butler now raises in that case.

Change pa1 metric for constrained photometry model.
It's not clear why pa1 increased in this case, but that will be investigated
in more detail as part of the meas_mosaic comparison test.