If visit id is not part of dataId, fetch it using butler
Current jointcal code requires a "visit" column to identify each
calexp. But "visit" may not be part of the dataId used for calexp,
depending on how the developer has chosen to organize directories.

This change still requires a visit id in the repository database, but
do not require that "visit" be one of the keys for "calexp". If it
is not, we assume that the visit column can nonetheless be fetched
using the butler and a dataId which is adequate for fetching a calexp.

Uses butler.queryMetadata(("visit"), dataId) to get the visit.