Add makeSkyWcs(), gtransfo.toAstMap(), test_gtransfo.py
makeSkyWcs() bypasses creation of a SIP and just stuffs all the gtransfos in a
model into an AST FrameDict.

BUGFIX! in inversePolyTransfo: the grid was not properly sampling the domain,
but was assuming it started at 0,0 and was recentering so it didn't include
the bounds of the box.

For inversePolyTransfo():
* More useful messages when there is a problem with the fit.
* Allow bailing when fitting if chi2 increases.
* Add TRACE logging for inversion steps.

Improve error reporting in getChipTransfo/getVisitTransfo

Expand gtransfo pybind11, adding astshim includes for toAstMap().
Support reading/writing polynomials to strings via python.
Increased stream output precision.