Add option to skip rank update when minimizing.
Add jointcalConfig option to skip photometry and astrometry rank update.
Add FitterBase support for skipping rank update.
Add decam photometry and astrometry tests with and without rank update.
Decam photometry doRankUpdate test doesn't work, and is part of the set of
"differs on macOS" tests, so we're skipping it for now.

Lift out Hessian creation into anonymous function.

NOTE: astrometryDoRankUpdate is mostly for testing: it shouldn't have any
measureable impact on the astrometry fitter, but it's helpful to demonstrate
that FitterBase.minimize() is behaving sanely when we doRankUpdate=False
(which is the point of the astrometry skip_rank_update test).