Adapt test for old config naming for calib flag
The data in testdata_jointcal used in this test was processed pre- RFC-498
(ticket DM-14997) and thus has the old flag names.  This sets the config in
the test to point to the old name appropriate for the current datasets in
testdata_jointcal.

This should be removed if the data in testdata_jointcal are ever reprocessed
post-RFC-498.