Move unit tests' retargeting of a.net to config file
Many unit tests in this package retarget LoadAstrometryNetObjectsTask
and then test JointcalTask with entry point `parseAndRun()`, the
CommandlineTask parser, which loads and applies the relevant
obs_package config overrides.
Move the astrometry and photometry RefObjLoader retargeting to a
config file so that it is applied last and the unit tests are not
affected by the obs_package configs.