Convert jointcal to use nanojansky
Use new utils Magnitude header for nJy/mag conversions in C++, and
astropy.units in python.

This adds a number of unfortunate hacks related to our refcats still being in
Jy (not nJy), but they are all called out with reference to RFC-549.

Moved the PhotometryModel validate() and checkPositiveOnBBox() tests into the
flux model, because it's a lot easier to make a flux-based model go negative.