Remove unused ref flux maps and filter lists
I had originally planned to pass lists of fluxes to the refcats for the future
calibration code, but they are just making things more clumsy, and it is not
clear whether that is even the behavior we would want in the future. Removing
them makes the code cleaner, they did nothing, and we can think through a
better API when we actually need it.

Update "bad" refStar code to reject `fluxErr <= 0`