use bigger index types
Extensive changes to the types used for indexing. Now using
Eigen::Index (defaults to std::ptrdiff_t) for indices and
std::size_t for counts. This requires using the cholmod_l_*
functions instead of cholmod_*. (A patch to the LSST Eigen
package handles the cholmod calls in there).

This is an effort to combat segfaults with large numbers of
HSC exposures (>~ 290).

Perhaps this is more extensive than it strictly needs to be
(e.g., polynomial orders don't need the full dynamic range of
std::size_t) but consistency helps avoid compiler warnings
and annoying casts.