Update DECam tests and configs for new data
Using Gaia-Dr2 lets us remove some config options.
These new numbers come from processing both single frame and jointcal
with Gaia for astrometry and PS1 for photometry.

I have removed the previously skipped tests because there are still fundamental
problems with the DECam data used here, and if we ever want to use better DECam
data we'll want to rewrite the tests from scratch. I added a disclaimer to the
top of the test class to note this.