Apply proper motion correction when reading refcats
Calculate epoch as the mean of all image dates; we could make this
configurable in the future.
Update test values: tests use Gaia, which has PM data, so the final fit
chi2 values change (most going down slightly, as expected). The
`dist_rms_absolute` values went up because the test comparison code
does not have the epoch to correct to, so that comparison will look
worse since the fit was against the corrected values.