Fix bugs in refcat epoch handling
Return the epoch used in a jointcal fit, and only compute it once for
both astrometry and photometry.
Use the returned epoch to load the test comparison refcat, as the gaia
refcat has requireProperMotion=True. With the DM-25807 cleanup of refcat
loading in meas_algorithms, not passing epoch when that option is set
results in an exception, as it always should have.