Update cfht_minimal test to gen3
Remove no_raise tests, as that is all handled internal to pipetask now;
keep the test of raising when there is no data since that is the
correct behavior now, always.

Add minSnr config for cfht_minimal tests.
Modify cfht_minimal-config to reflect the fact that the catalogs come from
much older processing.